requirements.txt
seleniumbase
fastapi
uvicorn[standard]
gunicorn
python-dotenv
thefuzz
mysql-connector-python